Reorganize submodule structure
