+++ lib/server.js
@@ -6,7 +6,7 @@ const metautil = require('metautil');
 const { Emitter, jsonParse, parseParams, split } = metautil;
 const { isError, receiveBody, delay, generateUUID } = metautil;
 
-const { WebsocketServer } = require('./websocket/server.js');
+const { WebsocketServer } = require('./websocket/ws.js');
 const { ServerTransport, buildHeaders } = require('./transport.js');
 const ServerHttpTransport = ServerTransport.transport.http;
 const ServerWsTransport = ServerTransport.transport.ws;
