+++ lib/websocket/connection.js
@@ -11,10 +11,7 @@ const {
   PROTOCOL_ERROR_SUBTYPES,
 } = require('./constants.js');
 const { Frame } = require('./frame.js');
-const { FrameParser } = require('./frameParser.js');
-const {
-  Utf8IncrementalValidator,
-} = require('./utils/utf8IncrementalValidator.js');
+const { FrameParser, isValidUTF8 } = require('./frameParser.js');
 
 class Connection extends EventEmitter {
   #socket;
@@ -27,7 +24,6 @@ class Connection extends EventEmitter {
   #closeReceived = false;
   #fragments = null;
   #closeTimer = null;
-  #utf8Validator = new Utf8IncrementalValidator();
 
   constructor(socket, head, options = {}) {
     super();
@@ -173,7 +169,7 @@ class Connection extends EventEmitter {
   }
 
   #handleDataFrame(frame) {
-    const { opcode, payload, fin } = frame;
+    const { opcode, payload } = frame;
     if (!this.#fragments) {
       // Continuation frame without a started fragmented message
       if (opcode === OPCODES.CONTINUATION) {
@@ -193,19 +189,7 @@ class Connection extends EventEmitter {
         const isBinary = opcode === OPCODES.BINARY;
         this.emit('message', frame.payload, isBinary);
       } else {
-        // First fragment
         if (!this.#trackMessageSize(payload.length)) return;
-        const isText = opcode === OPCODES.TEXT;
-        const isValid = this.#utf8Validator.push(payload, fin);
-        if (isText && !isValid) {
-          this.emit(
-            'error',
-            new Error('Invalid UTF-8 in text frame during fragments'),
-          );
-          return void this.#close(
-            Frame.errorClose('INVALID_PAYLOAD', this.#isClient),
-          );
-        }
         this.#fragments = {
           opcode,
           payloads: [payload],
@@ -216,23 +200,18 @@ class Connection extends EventEmitter {
       // continue fragments
       const totalSize = this.#fragments.totalSize + payload.length;
       if (!this.#trackMessageSize(totalSize)) return;
-      const isText = this.#fragments.opcode === OPCODES.TEXT;
-      const isValid = isText ? this.#utf8Validator.push(payload, fin) : true;
-      if (!isValid) {
-        this.emit(
-          'error',
-          new Error('Invalid UTF-8 in text frame during fragments'),
-        );
-        return void this.#close(
-          Frame.errorClose('INVALID_PAYLOAD', this.#isClient),
-        );
-      }
       this.#fragments.totalSize = totalSize;
       this.#fragments.payloads.push(frame.payload);
       if (!frame.fin) return;
       const isBinary = this.#fragments.opcode === OPCODES.BINARY;
+      const isText = this.#fragments.opcode === OPCODES.TEXT;
       const fullPayload = Buffer.concat(this.#fragments.payloads);
-      if (isText) this.#utf8Validator.reset();
+      if (isText && !isValidUTF8(fullPayload)) {
+        this.emit('error', new Error('Invalid UTF-8 in text frame'));
+        return void this.#close(
+          Frame.errorClose('INVALID_PAYLOAD', this.#isClient),
+        );
+      }
       this.#fragments = null;
       this.emit('message', fullPayload, isBinary);
     } else {
@@ -313,7 +292,6 @@ class Connection extends EventEmitter {
     this.#closing = true;
     this.#closeSent = true;
     this.#fragments = null;
-    this.#utf8Validator.reset();
 
     this.#socket.write(frameBuffer);
