+++ lib/websocket/frame.js
@@ -25,7 +25,7 @@ const {
   MAX_CLOSE_REASON,
   TWO_32,
 } = require('./constants.js');
-const { Result } = require('./utils/result.js');
+const { Result } = require('./result.js');
 
 const truncateCloseReason = (reason) => {
   const bytes = Buffer.from(reason, ENCODING);
