+++ lib/websocket/frameParser.js
@@ -17,7 +17,7 @@ const {
   TWO_32,
 } = require('./constants.js');
 const { Frame } = require('./frame.js');
-const { Result } = require('./utils/result.js');
+const { Result } = require('./result.js');
 
 const isValidUTF8 = (buf) => {
   let i = 0;
@@ -239,4 +239,5 @@ module.exports = {
   FrameParser,
   ParseError,
   PARSE_ERR_CODES,
+  isValidUTF8,
 };
