+++ lib/websocket/result.d.ts
