+++ lib/websocket/result.js
