+++ lib/websocket/ws.d.ts
@@ -1,0 +1,7 @@
+export * from './constants.js';
+export { WebsocketServer } from './server.js';
+export type { WebsocketServerOptions, VerifyClientInfo } from './server.js';
+export { Connection } from './connection.js';
+export type { ConnectionOptions } from './connection.js';
+export { Frame } from './frame.js';
+export { FrameParser, ParseError, PARSE_ERR_CODES } from './frameParser.js';
