+++ lib/websocket/ws.js
@@ -1,0 +1,21 @@
+'use strict';
+
+const constants = require('./constants.js');
+const { WebsocketServer } = require('./server.js');
+const { Connection } = require('./connection.js');
+const { Frame } = require('./frame.js');
+const {
+  FrameParser,
+  ParseError,
+  PARSE_ERR_CODES,
+} = require('./frameParser.js');
+
+module.exports = {
+  ...constants,
+  WebsocketServer,
+  Connection,
+  Frame,
+  FrameParser,
+  ParseError,
+  PARSE_ERR_CODES,
+};
