+++ metacom.js
@@ -3,5 +3,11 @@
 const { Metacom } = require('./lib/metacom.js');
 const { Server } = require('./lib/server.js');
 const { buildHeaders } = require('./lib/transport.js');
+const ws = require('./lib/websocket/ws.js');
 
-module.exports = { Metacom, Server, buildHeaders };
+module.exports = {
+  Metacom,
+  Server,
+  buildHeaders,
+  ...ws,
+};
